% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{File Read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file you would like R to read}
}
\value{
this function returns a data frame of the data present in the file
that was read in.
}
\description{
If it exists, reads in a file to be used for the other functions. The data
in this file will be presented as a data frame.
}
\note{
Will return an error if the file name does not exist in current working
directory.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read("accident_2014.csv.bz2")}
\dontrun{fars_read("accident_2015.csv.bz2")}

}
