% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Fars Read Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector of years you would like to take data from.}
}
\value{
Returns a data frame with two columns "MONTH" and "Year". Returns
the month and year for each row in the inputted data frame. If multiple years
are selected, you will be returned a data frame for each year individually.
}
\description{
reads in years from a data frame.
}
\note{
Will return an error if any of the given years do not contain any data.
}
\examples{
fars_read_years(2015)
fars_read_years(c(2014,2013,2015))
}
