% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fars Summary By Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector of years you want to summarize data from.}
}
\value{
This function returns a data frame with each column as a given year
in order of years given in the "years" parameter. Each row is representative
of a month in order from January to December.
}
\description{
Summarizes the number of fatalities from car crashes each year by month.
}
\note{
You will get an error if you input any years not contained in a data set.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014,2015))

}
