% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year that the data in the file is collected to be used as
part of the created filename.}
}
\value{
This function returns the filename that was generated from the function.
}
\description{
Creates a name for a file dependent on the year of the data.
}
\note{
all inputs into the parameter year will become integers, so make sure to
use the numerical year.
}
\examples{
\dontrun{make_filename(2013)}
\dontrun{make_filename(2014)}

}
